src/baliza/pyproject.toml
[tool.poetry]
name = "api-pncp-consulta-client"
version = "1.0"
description = "A client library for accessing API PNCP CONSULTA"
authors = []
readme = "README.md"
packages = [
    {include = "api_pncp_consulta_client"},
]
include = ["CHANGELOG.md", "api_pncp_consulta_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
